Lar gammel Typeahead komponent være, fjerner lokale testdata
